JavaConvention delegates all calls to JavaExtension
